refactor: fix typo in createTableBlockUISchema file
